Module aws_sdk_dataexchange::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

CancelJob

Fluent builder constructing a request to CancelJob.

CreateDataSet

Fluent builder constructing a request to CreateDataSet.

CreateEventAction

Fluent builder constructing a request to CreateEventAction.

CreateJob

Fluent builder constructing a request to CreateJob.

CreateRevision

Fluent builder constructing a request to CreateRevision.

DeleteAsset

Fluent builder constructing a request to DeleteAsset.

DeleteDataSet

Fluent builder constructing a request to DeleteDataSet.

DeleteEventAction

Fluent builder constructing a request to DeleteEventAction.

DeleteRevision

Fluent builder constructing a request to DeleteRevision.

GetAsset

Fluent builder constructing a request to GetAsset.

GetDataSet

Fluent builder constructing a request to GetDataSet.

GetEventAction

Fluent builder constructing a request to GetEventAction.

GetJob

Fluent builder constructing a request to GetJob.

GetRevision

Fluent builder constructing a request to GetRevision.

ListDataSetRevisions

Fluent builder constructing a request to ListDataSetRevisions.

ListDataSets

Fluent builder constructing a request to ListDataSets.

ListEventActions

Fluent builder constructing a request to ListEventActions.

ListJobs

Fluent builder constructing a request to ListJobs.

ListRevisionAssets

Fluent builder constructing a request to ListRevisionAssets.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

StartJob

Fluent builder constructing a request to StartJob.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateAsset

Fluent builder constructing a request to UpdateAsset.

UpdateDataSet

Fluent builder constructing a request to UpdateDataSet.

UpdateEventAction

Fluent builder constructing a request to UpdateEventAction.

UpdateRevision

Fluent builder constructing a request to UpdateRevision.