Module aws_sdk_dataexchange::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_limit_exceeded_exception

See ServiceLimitExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

Access to the resource is denied.

CancelJobError

ConflictException

The request couldn't be completed because it conflicted with the current state of the resource.

CreateDataSetError

CreateJobError

CreateRevisionError

DeleteAssetError

DeleteDataSetError

DeleteRevisionError

GetAssetError

GetDataSetError

GetJobError

GetRevisionError

InternalServerException

An exception occurred with the service.

ListDataSetRevisionsError

ListDataSetsError

ListJobsError

ListRevisionAssetsError

ListTagsForResourceError

ResourceNotFoundException

The resource couldn't be found.

ServiceLimitExceededException

The request has exceeded the quotas imposed by the service.

StartJobError

TagResourceError

ThrottlingException

The limit on the number of requests per second was exceeded.

UntagResourceError

UpdateAssetError

UpdateDataSetError

UpdateRevisionError

ValidationException

The request was invalid.

Enums

CancelJobErrorKind

CreateDataSetErrorKind

CreateJobErrorKind

CreateRevisionErrorKind

DeleteAssetErrorKind

DeleteDataSetErrorKind

DeleteRevisionErrorKind

GetAssetErrorKind

GetDataSetErrorKind

GetJobErrorKind

GetRevisionErrorKind

ListDataSetRevisionsErrorKind

ListDataSetsErrorKind

ListJobsErrorKind

ListRevisionAssetsErrorKind

ListTagsForResourceErrorKind

StartJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAssetErrorKind

UpdateDataSetErrorKind

UpdateRevisionErrorKind