Module aws_sdk_dataexchange::model

Modules

asset_destination_entry

See AssetDestinationEntry

asset_details

See AssetDetails

asset_entry

See AssetEntry

asset_source_entry

See AssetSourceEntry

data_set_entry

See DataSetEntry

details

See Details

export_asset_to_signed_url_request_details

See ExportAssetToSignedUrlRequestDetails

export_asset_to_signed_url_response_details

See ExportAssetToSignedUrlResponseDetails

export_assets_to_s3_request_details

See ExportAssetsToS3RequestDetails

export_assets_to_s3_response_details

See ExportAssetsToS3ResponseDetails

export_revisions_to_s3_request_details

See ExportRevisionsToS3RequestDetails

export_revisions_to_s3_response_details

See ExportRevisionsToS3ResponseDetails

export_server_side_encryption

See ExportServerSideEncryption

import_asset_from_signed_url_job_error_details

See ImportAssetFromSignedUrlJobErrorDetails

import_asset_from_signed_url_request_details

See ImportAssetFromSignedUrlRequestDetails

import_asset_from_signed_url_response_details

See ImportAssetFromSignedUrlResponseDetails

import_assets_from_s3_request_details

See ImportAssetsFromS3RequestDetails

import_assets_from_s3_response_details

See ImportAssetsFromS3ResponseDetails

job_entry

See JobEntry

job_error

See JobError

origin_details

See OriginDetails

request_details

See RequestDetails

response_details

See ResponseDetails

revision_destination_entry

See RevisionDestinationEntry

revision_entry

See RevisionEntry

s3_snapshot_asset

See S3SnapshotAsset

Structs

AssetDestinationEntry

The destination for the asset.

AssetDetails

AssetEntry

An asset in AWS Data Exchange is a piece of data that can be stored as an S3 object. The asset can be a structured data file, an image file, or some other data file. When you create an import job for your files, you create an asset in AWS Data Exchange for each of those files.

AssetSourceEntry

The source of the assets.

DataSetEntry

A data set is an AWS resource with one or more revisions.

Details

ExportAssetToSignedUrlRequestDetails

Details of the operation to be performed by the job.

ExportAssetToSignedUrlResponseDetails

The details of the export to signed URL response.

ExportAssetsToS3RequestDetails

Details of the operation to be performed by the job.

ExportAssetsToS3ResponseDetails

Details about the export to Amazon S3 response.

ExportRevisionsToS3RequestDetails

Details of the operation to be performed by the job.

ExportRevisionsToS3ResponseDetails

Details about the export revisions to Amazon S3 response.

ExportServerSideEncryption

Encryption configuration of the export job. Includes the encryption type as well as the AWS KMS key. The KMS key is only necessary if you chose the KMS encryption type.

ImportAssetFromSignedUrlJobErrorDetails

ImportAssetFromSignedUrlRequestDetails

Details of the operation to be performed by the job.

ImportAssetFromSignedUrlResponseDetails

The details in the response for an import request, including the signed URL and other information.

ImportAssetsFromS3RequestDetails

Details of the operation to be performed by the job.

ImportAssetsFromS3ResponseDetails

Details from an import from Amazon S3 response.

JobEntry

AWS Data Exchange Jobs are asynchronous import or export operations used to create or copy assets. A data set owner can both import and export as they see fit. Someone with an entitlement to a data set can only export. Jobs are deleted 90 days after they are created.

JobError

An error that occurred with the job request.

OriginDetails

RequestDetails

The details for the request.

ResponseDetails

Details for the response.

RevisionDestinationEntry

The destination where the assets in the revision will be exported.

RevisionEntry

A revision is a container for one or more assets.

S3SnapshotAsset

The S3 object that is the asset.

Enums

AssetType

The type of file your data is stored in. Currently, the supported asset type is S3_SNAPSHOT.

Code

JobErrorLimitName

The name of the limit that was reached.

JobErrorResourceTypes

The types of resource which the job error can apply to.

LimitName

Origin

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers). When an owned data set is published in a product, AWS Data Exchange creates a copy of the data set. Subscribers can access that copy of the data set as an entitled data set.

ResourceType

ServerSideEncryptionTypes

The types of encryption supported in export jobs to Amazon S3.

State

Type