Module aws_sdk_dataexchange::operation

Structs

CancelJob

This operation cancels a job. Jobs can be cancelled only when they are in the WAITING state.

CreateDataSet

This operation creates a data set.

CreateJob

This operation creates a job.

CreateRevision

This operation creates a revision for a data set.

DeleteAsset

This operation deletes an asset.

DeleteDataSet

This operation deletes a data set.

DeleteRevision

This operation deletes a revision.

GetAsset

This operation returns information about an asset.

GetDataSet

This operation returns information about a data set.

GetJob

This operation returns information about a job.

GetRevision

This operation returns information about a revision.

ListDataSetRevisions

This operation lists a data set's revisions sorted by CreatedAt in descending order.

ListDataSets

This operation lists your data sets. When listing by origin OWNED, results are sorted by CreatedAt in descending order. When listing by origin ENTITLED, there is no order and the maxResults parameter is ignored.

ListJobs

This operation lists your jobs sorted by CreatedAt in descending order.

ListRevisionAssets

This operation lists a revision's assets sorted alphabetically in descending order.

ListTagsForResource

This operation lists the tags on the resource.

StartJob

This operation starts a job.

TagResource

This operation tags a resource.

UntagResource

This operation removes one or more tags from a resource.

UpdateAsset

This operation updates an asset.

UpdateDataSet

This operation updates a data set.

UpdateRevision

This operation updates a revision.